test(no-extra-parens): cleanup duplicate test cases
